🧾 Rectify SPDX license identifier for `Apache-2.0`
The valid SPDX license identifier on https://spdx.org/licenses/ mentions that Apache-2.0 is the accepted standard instead of any other term. This is what I understand should be used for compliance with valid licensing specifications that might come into play later (such as when publishing a conda-forge package in #29).