🥌 Handle cross-compilation across platforms (#73)
This commit configures the `platform_tag` retrieval across all
three major platforms. This way, cross-compilation can be enabled
from any platform and architecture to any target that is supported
by the Zig compiler toolchain. Support for the Xcode toolchain has
been retained on macOS. Published wheels on PyPI will not be
cross-compiled since this is all still experimental and right now not
supported by the upstream Hugo project.

The experiments here confirm that with the correct choice of CC
and CXX and with GOOS and GOARCH, the wheels can be
cross-compiled for Windows (amd64/arm64/x86),
Linux (amd64/arm64), and macOS (amd64/arm64). It is to be
noted that CIBUILDWHEEL restricts the cross-compilation on
macOS and for good reason, since we want to compile natively.

Future work here can improve the error handling to detect the
various compiler toolchains present.